Fix map sharing for jsonmap and indexedmap and test case for them.

+++ core/main/java/com/codingchili/core/testing/MapTestCases.java
@@ -748,7 +748,7 @@ public class MapTestCases {
                     store.size(size -> {
                         newStorage.size(newSize -> {
                             test.assertEquals(size.result(), newSize.result());
-                            test.assertEquals(newSize.result(), TEST_ITEM_COUNT);
+                            test.assertEquals(newSize.result(), TEST_ITEM_COUNT.intValue());
                             async.complete();
                         });
                     });
